+++ src/components/mobile/RenderMobile.tsx
@@ -109,7 +109,7 @@ const RenderMobile: React.FC<RenderMobile> = ({
   return (
     <div className='col-4'>
       <div ref={div}>
-        <CardHeaderPromotion textHeader='ตัวอย่างในแอปพลิเคชั่น' center={true} />
+        <CardHeaderPromotion textHeader='ตัวอย่างในแอปพลิเคชัน' center={true} />
         <div
           style={{
             width: '100%',