+++ src/pages/highlightNews/HighlightNewsPage.tsx
@@ -118,7 +118,7 @@ function HighlightNewsPage() {
           <div className='col-lg'>
             <Select
               className='col-lg-12'
-              placeholder='เลือกแอปพลิเคชั่น'
+              placeholder='เลือกแอปพลิเคชัน'
               onChange={onChangeApplication}
               showSearch
               value={application}
@@ -301,7 +301,7 @@ function HighlightNewsPage() {
       key: 'name',
     },
     {
-      title: 'แอปพลิเคชั่น',
+      title: 'แอปพลิเคชัน',
       dataIndex: 'application',
       key: 'application',
       render: (value: any, row: any, index: number) => {