+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [Version - 0]
 
+## Fix [v0.41.3] 2023-12-12
+
+- fixed logic of add / delete images
+
+## improve [v0.41.2] 2023-12-10
+
+- edit image in finish task
+
 ## Fix [v0.41.1] 2023-12-1
 
 - fixed date time in edit guru kaset page